Change month from short to long

+++ js/date.js
@@ -1,2 +1,2 @@
-const d = new Date().toLocaleDateString('en-us', { weekday:"long", year:"numeric", month:"short", day:"numeric"}) 
+const d = new Date().toLocaleDateString('en-us', { weekday:"long", year:"numeric", month:"long", day:"numeric"}) 
 $('h1').text(d);
